requirements.txt
streamlit
openai
pandas
python-dotenv
tiktoken
textblob
